backend/requirements.txt
# SmartSecure Sri Lanka Backend Dependencies - Phase 3
# Core Flask Framework
Flask==3.1.2
flask-cors==4.0.1

# Database & ORM
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.23
psycopg2-binary==2.9.7

# Security & Authentication
PyJWT==2.10.1
bcrypt==5.0.0
flask-limiter==3.13

# AI & Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.4

# Environment & Configuration
python-dotenv==1.1.1

# Additional Utilities
Werkzeug==3.1.3

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0